Refactor to follow platform SDK styling
